Feature/config granularity (#15)
* Load intended configuration explicitly, return default correctly.

This decouples the call order from the test succeeding, and affords
the assumption that retrieving items from the configuration
is non mutating, and will give the same results.

* Refactor tests into reusable stages

* Add max-age, subdomains, and preload config support for hsts

* Add tests for hsts

* PHP7 compat

This is identical in function (`null` default adds nullability), but
is compatable with PHP7, unlike the explicit nullable notation.

* Match existing camelCase convention in config string

* Format comments

* Add strict mode

* Add CSP tests and CSPRO

* Rework API for applySecureHeadersWithConfig (again)

* More assertions on expect-ct

Check that header is present, check standard headers are present

* Change hsts to require maxAge or enabled in order to sent the header
(i.e. enabled no longer required if maxage specified).